Implemented the passing pen with basic adoption within the replay mode.
Allowed pass-candidates to be wrapped in a dynamically rendered anchor element for enabling clicking and hovering. Utility functions included for adding/removing pass-candidacy to effectively change between pass- and non-pass edit modes. Making the pass updates the CSS representation and the nextState as denoted by the current transition. Replay mode supports the pass superficially through a colour change, but an actual `visual pass' would need to be implemented shortly.