Added support for zone-formation `templates' in play setup.
Uploaded a fixed set of states in the back-end as JSONs. In the set-up stage, user can choose to either drag to rearrange a formation or apply one of the `built-in' ones. Allows for extensibility since uploading to the back-end server will update the list of available formations to choose from via the API get request.